+++ config/__init__.py
@@ -1,1 +1,1 @@
-from config.config import read_config_option, write_config_option, setup
+from config.config import setup, read_config_option, set_config_options, write_config_options_to_file, set_config_file